tests/python/requirements.txt
playwright==1.32.1
pytest==7.3.1
pytest-playwright==0.3.0
python-dotenv==1.0.0